transferring funds between wallets using convertion-rates-mechanism

+++ wallets/migrations/01-initial.sql
@@ -5,6 +5,6 @@ USE `wallets`;
 CREATE TABLE IF NOT EXISTS `wallets` (
     `Login` varchar(255) NOT NULL,
     `BaseCurrency` tinyint(4) NOT NULL,
-    `Balance` int(11) DEFAULT 0,
+    `Balance` decimal(16,4) DEFAULT 0.0,
     PRIMARY KEY (`Login`, `BaseCurrency`)
 ) ENGINE=InnoDB DEFAULT CHARSET=utf8;
